Add files via upload
This is a modlet that adds a bunch of different dye colors for 7 Days To Die.
In this lastest update I have added recipes for each dye color including the vanilla colors.
This modlet should work in any overhaul mod in SP and MP.
More is yet to come so stick around to see what is added as well as other projects.
If you have a color request you can contact either myself DragonWolf1982#8894 or Unforgiven774#2535 on Discord.
I couldn't have done it without the help of my friends Unforgiven774 and Khaine himself (creator of Darkness Falls overhaul mod). Thank you so much my friends.
I do hope you all enjoy this modlet.